+++ r-aks.tf
@@ -79,7 +79,7 @@ resource "azurerm_kubernetes_cluster" "aks" {
   }
 
   dynamic "aci_connector_linux" {
-    for_each = var.aci_subnet_id != null && var.aks_network_plugin != "kubenet" ? [true] : []
+    for_each = length(var.aci_subnet_id) > 0 && var.aks_network_plugin != "kubenet" ? [true] : []
     content {
       subnet_name = element(split("/", var.aci_subnet_id), length(split("/", var.aci_subnet_id)) - 1)
     }
@@ -161,7 +161,7 @@ resource "azurerm_kubernetes_cluster" "aks" {
       error_message = "var.pods_subnet must be set when using Azure CNI Cilium network."
     }
     precondition {
-      condition     = try(jsondecode(data.azapi_resource.subnet_delegation.output).properties.delegations[0].properties.serviceName, null) == "Microsoft.ContainerInstance/containerGroups"
+      condition     = try(jsondecode(data.azapi_resource.subnet_delegation[0].output).properties.delegations[0].properties.serviceName, null) == "Microsoft.ContainerInstance/containerGroups"
       error_message = "ACI subnet should be delegated to Microsoft.ContainerInstance/containerGroups"
     }
   }